Resolved deoendency issues and restructured folders
